scripts/role-assignment-4000-limit/requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-authorization
azure-graphrbac
msgraph-core
psutil
requests
pandas
openpyxl